Updated weapon + enemy(?)
Updated the weapon mechanics so the pistol alt fire is now working (the hurtboxes and animations change when it times out or collides with stuff) and the primary fire can collide with things (added a pierce variable for possible future content)

Added an enemy that follows the player when the player enters a detection zone.
There is currently a bug with the enemy where it will always detect the player on startup and it will sometimes lose the player after following for some time.